학습 노트 생성 기능 관련 UX 수정

+++ src/components/posts/PostLIstItem.vue
@@ -22,7 +22,10 @@ export default {
   },
   methods: {
     async deleteItem() {
-      await deletePost(this.postItem._id);
+      if (confirm('You want to delete it?')) {
+        await deletePost(this.postItem._id);
+        this.$emit('refresh');
+      }
     },
   },
 };
